+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Added
 - "payment" account type for Pix related resources
 - missing parameters to Boleto, BrcodePayment, Deposit, DictKey, Event, Invoice, Transfer and Workspace resources
+- Workspace.update() to allow parameter updates
 
 ### Fixed
 - special characters in brcodePreview query